+++ board/safety.h
@@ -13,6 +13,7 @@
 #include "safety/safety_mazda.h"
 #include "safety/safety_nissan.h"
 #include "safety/safety_volkswagen_mqb.h"
+#include "safety/safety_volkswagen_pq.h"
 #include "safety/safety_elm327.h"
 
 // from cereal.car.CarParams.SafetyModel
